refactor(services): migrate small files to subpackages (1/2)
Move event_helpers.py and task_binding_guard.py:
- event_helpers.py (22 lines) → shared/events.py
- task_binding_guard.py (42 lines) → task/binding_guard.py

Both files have clear responsibilities and are used across the codebase:
- emit_issue_failed: avoids circular dependency in roles layer
- task binding guard: validation logic for PR creation

Updated all imports and lazy import mappings in:
- services/__init__.py
- shared/__init__.py
- task/__init__.py
- pr/create.py
