+++ src/vibe3/services/shared/events.py
